requirements.txt
beautifulsoup4==4.4.1
requests==2.8.0
dnspython==1.15.0